Handle error on the UDP socket.
When some error occurs on the UDP socket (e.g. ack is
send to the peer which already close its socket), runtime
sends a message {udp_error, Socket, Reason}.

This message must be handled (otherwise procress fails on
handle_info functin clause error). This is similar like
tcp_error message, but it is not mentioned in the documentation.